ad feature tarin new images

+++ train_models.py
@@ -7,8 +7,11 @@ from feature_extractor import FeatureExtractor
 fe = FeatureExtractor()
 
 for img_path in sorted(glob.glob('static/img/*.jpg')):
-    print(img_path)
-    img = Image.open(img_path)
-    feature = fe.extract(img)
-    feature_path = 'static/feature/' + os.path.splitext(os.path.basename(img_path))[0] + '.pkl'
-    pickle.dump(feature, open(feature_path, 'wb'))
+    if os.path.isfile('static/feature/' + os.path.splitext(os.path.basename(img_path))[0] + '.pkl'):
+        print(img_path, "was trained befor...")
+    else:
+        print("NEW IMAGE TRAINED:", img_path)
+        img = Image.open(img_path)
+        feature = fe.extract(img)
+        feature_path = 'static/feature/' + os.path.splitext(os.path.basename(img_path))[0] + '.pkl'
+        pickle.dump(feature, open(feature_path, 'wb'))
